requirements.txt
# ── Web Framework ──────────────────────────────
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.30.0,<1.0
httpx>=0.27.0,<1.0

# ── LangChain + Ollama ────────────────────────────
langchain>=0.3.7,<0.4
langchain-community>=0.3.7,<0.4
langchain-ollama>=0.2.0,<0.3
langchain-text-splitters>=0.3.2,<0.4
langchain-chroma>=0.2.0,<0.3

# ── Vector DB ─────────────────────────────────────
chromadb>=0.5.5,<0.6

# ── PDF Parsing ───────────────────────────────────
pypdf>=4.3.0,<5.0

# ── Config / Utils ────────────────────────────────
pydantic>=2.5,<3.0
pydantic-settings>=2.5,<3.0
python-dotenv>=1.0,<2.0

# ── Ollama Client ──────────────────────────────────
ollama>=0.6.0,<1.0

# ── Oracle 23ai (Analytics DB) ────────────────────
oracledb>=2.0,<3.0
